@@ -24,6 +24,7 @@ import com.google.gson.reflect.TypeToken;
 import com.sk89q.worldedit.util.net.HttpRequest;
 
 import java.io.IOException;
+import java.net.URI;
 import java.net.URL;
 import java.util.Map;
 import java.util.concurrent.Callable;
@@ -33,23 +34,38 @@ public class EngineHubPaste implements Paster {
     private static final Gson GSON = new Gson();
 
     @Override
-    public Callable<URL> paste(String content) {
-        return new PasteTask(content);
+    public Callable<URL> paste(String content, PasteMetadata metadata) {
+        return new PasteTask(content, metadata);
     }
 
     private static final class PasteTask implements Callable<URL> {
 
         private final String content;
+        private final PasteMetadata metadata;
 
-        private PasteTask(String content) {
+        private PasteTask(String content, PasteMetadata metadata) {
             this.content = content;
+            this.metadata = metadata;
         }
 
         @Override
         public URL call() throws IOException, InterruptedException {
             URL initialUrl = HttpRequest.url("https://paste.enginehub.org/signed_paste");
 
-            SignedPasteResponse response = GSON.fromJson(HttpRequest.get(initialUrl)
+            HttpRequest requestBuilder = HttpRequest.get(initialUrl);
+
+            requestBuilder.header("x-paste-meta-from", "EngineHub");
+            if (metadata.name != null) {
+                requestBuilder.header("x-paste-meta-name", metadata.name);
+            }
+            if (metadata.author != null) {
+                requestBuilder.header("x-paste-meta-author", metadata.author);
+            }
+            if (metadata.extension != null) {
+                requestBuilder.header("x-paste-meta-extension", metadata.extension);
+            }
+
+            SignedPasteResponse response = GSON.fromJson(requestBuilder
                     .execute()
                     .expectResponseCode(200)
                     .returnContent()
@@ -68,7 +84,7 @@ public class EngineHubPaste implements Paster {
                     .execute()
                     .expectResponseCode(200, 204);
 
-            return new URL(response.viewUrl);
+            return URI.create(response.viewUrl).toURL();
         }
 
     }
